+++ packages/upload-api/src/blob/allocate.js
@@ -12,6 +12,7 @@ export function blobAllocateProvider(context) {
     W3sBlob.allocate,
     async ({ capability, invocation }) => {
       const { blob, cause, space } = capability.nb
+      let size = blob.size
 
       // Rate limiting validation
       // TODO: we should not produce rate limit error but rather suspend / queue task to be run after enforcing a limit without erroring
@@ -55,17 +56,30 @@ export function blobAllocateProvider(context) {
         // added to the space and there is no allocation change.
         // If record exists but is expired, it can be re-written
         if (allocationInsert.error.name === 'RecordKeyConflict') {
+          size = 0
+        } else {
           return {
-            ok: { size: 0 },
+            error: allocationInsert.error,
           }
         }
+      }
+
+      // Check if blob already exists
+      const hasBlobStore = await context.blobsStorage.has(blob.digest)
+      if (hasBlobStore.error) {
+        return hasBlobStore
+      }
+
+      // If blob is stored, we can just allocate it to the space with the allocated size
+      if (hasBlobStore.ok) {
         return {
-          error: allocationInsert.error,
+          ok: { size },
         }
       }
 
       // Get presigned URL for the write target
       const expiresIn = 60 * 60 * 24 // 1 day
+      const expiresAt = (new Date(Date.now() + expiresIn)).toISOString()
       const createUploadUrl = await context.blobsStorage.createUploadUrl(
         blob.digest,
         blob.size,
@@ -75,27 +89,15 @@ export function blobAllocateProvider(context) {
         return createUploadUrl
       }
 
-      // Check if blob already exists
-      const hasBlobStore = await context.blobsStorage.has(blob.digest)
-      if (hasBlobStore.error) {
-        return hasBlobStore
-      }
-
-      // If blob is stored, we can just allocate it to the space
-      if (hasBlobStore.ok) {
-        return {
-          ok: { size: blob.size },
-        }
-      }
-
       const address = {
         url: createUploadUrl.ok.url.toString(),
         headers: createUploadUrl.ok.headers,
+        expiresAt
       }
 
       return {
         ok: {
-          size: blob.size,
+          size,
           address,
         },
       }
