+++ packages/upload-api/test/handlers/web3.storage.spec.js
@@ -1,0 +1,4 @@
+import { test } from '../test.js'
+import * as W3s from './web3.storage.js'
+
+test({ 'web3.storage/*': W3s.test })
